Xresources, awesome: Switched distro
I decided to abandon Arch and install something else: Linux Mint. Mint
has a desktop environment called MATE, which is basically a fork of
GNOME2. I replaced the default window manager (marco) with awesome in
order to get access to all that sweet dynamic window management. The
startup is slightly longer now that there are tons of GNOME services
starting up on login, but atleast they work pretty much out-of-the-box.

Xresources:
- Made white more white
- Changed xterm font and urxvt url launcher.

awesome:
- Back to using urxvt. For some reason the Mint version of xterm
  couldn't display wide characters.
- Replaced old programs with MATE programs.
- Added programs for controlling MATE session.
- Now using a more robust script for controlling volume.
- XF86AudioPrevious is not a valid key.
- For some reason Firefox and Spotify kept launching in floating
  mode. Maybe MATE has something to do with this? I dunno.
- Changed theme font from profont to terminus.